+++ src/nvim/syntax.c
@@ -6593,7 +6593,7 @@ do_highlight (
            * Copy characters from arg[] to buf[], translating <> codes.
            */
           for (p = arg, off = 0; off < 100 - 6 && *p; ) {
-            len = trans_special(&p, buf + off, FALSE);
+            len = (int)trans_special(&p, buf + off, FALSE);
             if (len > 0)                    /* recognized special char */
               off += len;
             else                            /* copy as normal char */
